Fix "This Minute" time filter: parseInt→parseFloat for sub-hour values
parseInt("0.0167") returned 0, making the "This Minute" filter behave
like "All Time". parseFloat correctly handles the fractional hour value
while still working for all integer hour values.

+++ static/js/app.js
@@ -1514,7 +1514,7 @@
   }
   function computeFilteredState() {
     const searchText = (document.getElementById("search-box").value || "").trim().toLowerCase();
-    const timeHours = parseInt(document.getElementById("filter-time").value) || 0;
+    const timeHours = parseFloat(document.getElementById("filter-time").value) || 0;
     const hideOpinion = document.getElementById("filter-opinion")?.checked || false;
     const now = Date.now();
     const originFiltering = state.activeOrigins.size < 15;
